net: e1000: add read_rom_hwaddr eth uclass ops
read_rom_hwaddr is a optional eth uclass ops, which should
be implemented by devices which have MAC address stored in
ROM on the board, to expose it to the network stack.
e1000 driver did not have it, since it is optional, and it
exposes its own MAC address to platdata in probe routine.
This patch adds read_rom_hwaddr ops, so eth uclass driver
could know that it has its own MAC, and would not overwritten
it with u-boot ethaddr env.

Change-Id: Id49b5f4b0a9bf83c4bc2e1f13cbaf381c38993e6
Signed-off-by: jinghua <jinghua@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/43755
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>