usb: xhci: Don't assume LS/FS devices are always behind a HS hub
At present xHCI driver assumes LS/FS devices are attached directly
to a HS hub. If they are connected to a LS/FS hub, the driver will
fail to perform the USB enumeration process on such devices.

This is fixed by looking from the device itself all the way up to
the HS hub where the TT that serves the device is located.

Change-Id: I0b0fae3b5aad8ff1d81d07229bde6173389f7b3a
Signed-off-by: Bin Meng <bmeng.cn@gmail.com>