usb: Read device descriptor after device is addressed for xHCI
For xHCI it is not possible to read a device descriptor before it
has been assigned an address. That's why usb_setup_descriptor()
was called with 'do_read' being false. But we really need try to
read the device descriptor before starting any real communication
with the default control endpoint.

Change-Id: I4eb3d37386fa433bc424e8b785d979d3e03958e5
Signed-off-by: Bin Meng <bmeng.cn@gmail.com>