ehci: mvebu: toggle vbus for EHCI start and stop
Currently, the VBUS for EHCI is fixed on 5v and never toggle it
when do usb start and usb stop.
According the USB2.0 SPEC, the timing between VBUS on and USB
enumeration should be 200ms. But now the timing depends on
when to run the command "usb start" and it maybe a long time. It
is not follow the USB2.0 SPEC and will cause some u-disk misdetection.
This patch add toggle VBUS for EHCI to follow the time requirement
in USB2.0 SPEC and also add marvell.ehci-usb document.

Change-Id: Ia6acc0f80a7344b1fc650e567176011ddd802305
Signed-off-by: allen yan <yanwei@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/47434
Reviewed-by: Igal Liberman <igall@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Reviewed-by: David Sniatkiwicz <davidsn@marvell.com>