mvebu: pcie: dw: add correct ATU handling
Currently PCIe is not functional on Armada8kp because of
incorrect handling of ATU (address translation unit) for
addresses bigger than 32bits which is the case in Armada8kp.

Currently, ATU is configured once, during the probe of the driver.
However, we need to configure ATU differently for different
accesses (memory, configuration and IO).

This patch uses the same approach as used in Linux Kernel.

Change-Id: I162b3e5812b1ce69389d6e27afe3af61c8aa4c0f
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/50543
Tested-by: iSoC Platform CI <ykjenk@marvell.com>