fix: spi: set spi speed to the minimum between controller's and flash's frequency
Both spi controller and flash have got the maximum working freq, the
correct freq should be the lower one of these two.
Now there is a bug that when flash's freq is higher than controller's,
spi uclass sets the working freq to flash's, instead of controller's.
This patch sets spi speed to the minimum between controller's and
flash's frequency.

Change-Id: If337f65a0be2146cc15279810c8edefb2a2eb7d2
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/53207
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>