spi: set spi speed to the minimum between controller's and flash's frequency
Both spi controller and flash have got the maximum working freq, the
correct freq should be the lower one of these two.
Now there is a bug that when flash's freq is higher than controller's,
spi uclass sets the working freq to flash's, instead of controller's.
This patch sets spi speed to the minimum between controller's and
flash's frequency.

Change-Id: Ibb9c54795cd7022fd0f8106f0c16cb976fd97266
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>