a8k: set cpuidle.off=1 in bootargs for a8040 A1/2
Linux kernel cpuidle feature is not supported
in Armada8040 A1/A2 revisions, in order to be able
to enable it by default in Linux kernel, we need
to set in Linux kernel bootargs cpuidle.off=1.

This patch adds $cpuidle variable to set_bootargs.
By default $cpuidle is empty.
For SoC revision is A1/2, set $cpuidle to cpuidle.off=1,
this will disable cpuidle in Linux kernel.

In addition, enable CONFIG_ARCH_MISC_INIT, in order
to use arch_misc_init() for this purpose.

Change-Id: I3f742aa213f84760be4ddff2b175bb5b7eea571d
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/56770
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/57189