a3900: enable SSCG feature
Spread Spectrum Clock Generator (SSCG) used to reduce system EMI emission.
The SSCG is applied to the DDR interface clock and the DDR clock.
The standard values for these parameters are a modulation frequency
of 30 kHz and a deviation of 1%.
This patch enables SSCG feature on Armada-3900-A1 with
DDR frequencies 800MHz and 1200MHz.

Change-Id: I1d417f46ad8fd1049af27837cde3f840eb3da1e4
Signed-off-by: David Sniatkiwicz <davidsn@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/57309
Reviewed-by: Igal Liberman <igall@marvell.com>
Tested-by: Igal Liberman <igall@marvell.com>