board: octeontx2 octeontx: bootargs without debug
By default the "debug" should not present in bootargs and in
Kernel-command-line.

The "debug" presence may cause a pollution of un-wanted debug-prints and
therefore should not be present/enabled by default, but only manually on
purpose by a sw designer.

One of pollution's examples is
systemd-journald[3314]: Journal effective settings seal=no compress=no \
   compress_threshold_bytes=512B

Change-Id: I701b0cf5968925ecae359e10fd553ab1c8ea4a28
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/13363
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Suneel Garapati <sgarapati@caviumnetworks.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>