xhci: octeontx2: Work-around designware USB issue
In the Designware XHCI core used in the OcteonTX2
SoC there is a bug if the received data is short
and it crosses a 64K boundary.  In this case,
instead of a single TRB event, two events are
generated.  The first event is Interrupt-on
Short Packet (ISP) and the next event is
Interrupt on Completion (IOC).  Also, in this
case, both completion events contain the wrong
length which does not make any sense.
In order to work around this ugly bug, a 64K
bounce buffer that is aligned to a 64K boundary
is used and the data is later copied to the
actual buffer.  This is only used if the entire
transaction is no greater than 64K.  So far this
issue has not manifested itself in larger
transactions.

Change-Id: I9228ef9e7c94f5cfefc8e905af433f328355d1fa
Signed-off-by: Aaron Williams <aaron.williams@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/13426
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>