net: tftp: Skip transmit of extra ack packet
With TFTPPUT support enabled and reading a file,
data-ack sequence with windowsize 1, an additional ack
is being sent for the last packet causing server to respond
with ICMP error.
This change is introduced in
commit "cc6b87ecaa96": net: tftp: Add client support for RFC 7440.

For now skip the extra packet send as it will not effect default
windowsize and helps to fix failure with file read.

Need re-visit with proper fix later.

Change-Id: If1838c9215506dd2198e9460fbe89c6d44917763
Signed-off-by: Suneel Garapati <sgarapati@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/43615
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>